requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
liboqs-python>=0.10.1
cryptography>=42.0.0
pydantic>=2.7.0
requests>=2.31.0
